requirements.txt
matplotlib
numpy
numpy-stl
scikit-image
